PR feedback: renaming + additional scroll for both directions
