+++ js/renovation/ui/scroll_view/scrollbar_props.ts
@@ -11,4 +11,6 @@ export class ScrollBarProps extends ScrollableInternalProps {
   @OneWay() showScrollbar?: ScrollableShowScrollbar;
 
   @OneWay() direction: ScrollableDirection = 'vertical';
+
+  @OneWay() scrollByThumb = false;
 }
